refactor: move base.ts to core folder and update imports
